django/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-celery-beat==2.4.0
django-debug-toolbar==3.8.1

django-filter==22.1
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
kombu==5.2.4
Markdown==3.4.1
pep8==1.7.1
Pillow==9.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
Pygments==2.14.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
